move message/mod.rs into separate files
