+++ chain-signatures/node/src/protocol/message/crypto.rs
@@ -1,0 +1,384 @@
+//! Encryption and wire serialization for peer messages.
+
+use crate::protocol::contract::primitives::ParticipantMap;
+
+use crate::protocol::message::types::MessageError;
+
+use cait_sith::protocol::Participant;
+use mpc_keys::hpke::{self, Ciphered};
+use near_crypto::Signature;
+use serde::de::DeserializeOwned;
+use serde::{Deserialize, Serialize};
+
+/// A signed message that can be encrypted. Note that the message's signature is included
+/// in the encrypted message to avoid from it being tampered with without first decrypting.
+#[derive(Serialize, Deserialize)]
+pub struct SignedMessage {
+    /// The message with all it's related info.
+    #[serde(with = "serde_bytes")]
+    pub msg: Vec<u8>,
+    /// The signature used to verify the authenticity of the encrypted message.
+    pub sig: Signature,
+    /// From which particpant the message was sent.
+    pub from: Participant,
+}
+
+impl SignedMessage {
+    pub const ASSOCIATED_DATA: &'static [u8] = b"";
+}
+
+impl SignedMessage {
+    pub fn encrypt<T: Serialize>(
+        msg: &T,
+        from: Participant,
+        sign_sk: &near_crypto::SecretKey,
+        cipher_pk: &hpke::PublicKey,
+    ) -> Result<Ciphered, MessageError> {
+        let msg = cbor_to_bytes(msg)?;
+        let sig = sign_sk.sign(&msg);
+        let msg = Self { msg, sig, from };
+        let msg = cbor_to_bytes(&msg)?;
+        let ciphered = cipher_pk
+            .encrypt(&msg, Self::ASSOCIATED_DATA)
+            .inspect_err(|err| {
+                tracing::error!(?err, "failed to encrypt message");
+            })?;
+        Ok(ciphered)
+    }
+}
+
+impl SignedMessage {
+    pub fn decrypt<T: DeserializeOwned>(
+        encrypted: &Ciphered,
+        cipher_sk: &hpke::SecretKey,
+        participants: &ParticipantMap,
+    ) -> Result<T, MessageError> {
+        Self::decrypt_with(encrypted, cipher_sk, participants, |_| Ok(()))
+    }
+
+    pub fn decrypt_with<T: DeserializeOwned, F: FnMut(&Signature) -> Result<(), MessageError>>(
+        encrypted: &Ciphered,
+        cipher_sk: &hpke::SecretKey,
+        participants: &ParticipantMap,
+        mut check: F,
+    ) -> Result<T, MessageError> {
+        let msg = cipher_sk
+            .decrypt(encrypted, Self::ASSOCIATED_DATA)
+            .inspect_err(|err| {
+                tracing::error!(?err, "failed to decrypt message");
+            })?;
+        let Self { msg, sig, from } = cbor_from_bytes(&msg)?;
+        let info = participants
+            .get(&from)
+            .ok_or(MessageError::UnknownParticipant(from))?;
+
+        // Do external check before verifying the signature.
+        check(&sig)?;
+
+        if !sig.verify(&msg, &info.sign_pk) {
+            tracing::error!(?from, "signed message erred out with invalid signature");
+            return Err(MessageError::Verification(
+                "invalid signature while verifying authenticity of encrypted protocol message",
+            ));
+        }
+
+        cbor_from_bytes(&msg)
+    }
+}
+
+pub fn cbor_to_bytes<T: Serialize + ?Sized>(value: &T) -> Result<Vec<u8>, MessageError> {
+    let mut buf = Vec::new();
+    ciborium::into_writer(value, &mut buf)
+        .map_err(|err| MessageError::CborConversion(err.to_string()))?;
+    Ok(buf)
+}
+
+pub(crate) fn cbor_from_bytes<T: DeserializeOwned>(bytes: &[u8]) -> Result<T, MessageError> {
+    ciborium::from_reader(bytes).map_err(|err| MessageError::CborConversion(err.to_string()))
+}
+
+pub(crate) const fn cbor_name(value: &ciborium::Value) -> &'static str {
+    match value {
+        ciborium::Value::Integer(_) => "integer",
+        ciborium::Value::Bytes(_) => "bytes",
+        ciborium::Value::Text(_) => "text",
+        ciborium::Value::Float(_) => "float",
+        ciborium::Value::Null => "null",
+        ciborium::Value::Bool(_) => "bool",
+        ciborium::Value::Array(_) => "array",
+        ciborium::Value::Map(_) => "map",
+        ciborium::Value::Tag(_, _) => "tag",
+        _ => "unknown",
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::protocol::contract::primitives::{ParticipantMap, Participants};
+    use crate::protocol::message::{GeneratingMessage, Message, SignatureMessage, TripleMessage};
+    use crate::protocol::ParticipantInfo;
+
+    use cait_sith::protocol::Participant;
+    use mpc_keys::hpke::{self, Ciphered};
+    use mpc_primitives::SignId;
+    use serde::{de::DeserializeOwned, Deserialize, Serialize};
+
+    #[test]
+    fn test_sending_encrypted_message() {
+        let associated_data = b"";
+        let (cipher_sk, cipher_pk) = mpc_keys::hpke::generate();
+        let starting_message = Message::Generating(GeneratingMessage {
+            from: cait_sith::protocol::Participant::from(0),
+            data: vec![],
+        });
+
+        let message = serde_json::to_vec(&starting_message).unwrap();
+        let message = cipher_pk.encrypt(&message, associated_data).unwrap();
+
+        let message = serde_json::to_vec(&message).unwrap();
+        let cipher = serde_json::from_slice(&message).unwrap();
+        let message = cipher_sk.decrypt(&cipher, associated_data).unwrap();
+        let message: Message = serde_json::from_slice(&message).unwrap();
+
+        assert_eq!(starting_message, message);
+    }
+
+    #[test]
+    fn test_encrypt_then_decrypt() {
+        let (cipher_sk, cipher_pk) = mpc_keys::hpke::generate();
+        let sign_sk =
+            near_crypto::SecretKey::from_seed(near_crypto::KeyType::ED25519, "sign-encrypt0");
+        let from = Participant::from(7);
+        let mut participants = Participants::default();
+        participants.insert(
+            &from,
+            ParticipantInfo {
+                sign_pk: sign_sk.public_key(),
+                cipher_pk: cipher_pk.clone(),
+                id: from.into(),
+                url: "http://localhost:3030".to_string(),
+                account_id: "test.near".parse().unwrap(),
+            },
+        );
+        let participants = ParticipantMap::One(participants);
+
+        let batch = vec![Message::Triple(TripleMessage {
+            id: 1234,
+            epoch: 0,
+            from,
+            data: vec![128u8; 1024],
+            timestamp: 1234567,
+        })];
+        let encrypted = SignedMessage::encrypt(&batch, from, &sign_sk, &cipher_pk).unwrap();
+        let decrypted_batch: Vec<Message> =
+            SignedMessage::decrypt(&encrypted, &cipher_sk, &participants).unwrap();
+
+        assert_eq!(
+            batch, decrypted_batch,
+            "batch messages did not get encrypted and decrypted correctly"
+        );
+    }
+
+    #[test]
+    fn test_serialization_change() {
+        #[derive(Serialize, Deserialize)]
+        struct NewSignedMessage {
+            #[serde(with = "serde_bytes")]
+            msg: Vec<u8>,
+            sig: near_crypto::Signature,
+            from: Participant,
+
+            // default will call Default::default() if missing in serialized bytes.
+            #[serde(default)]
+            added_field: Vec<u32>,
+        }
+
+        impl NewSignedMessage {
+            const ASSOCIATED_DATA: &'static [u8] = SignedMessage::ASSOCIATED_DATA;
+
+            fn encrypt<T: Serialize>(
+                batch: &T,
+                from: Participant,
+                sign_sk: &near_crypto::SecretKey,
+                cipher_pk: &hpke::PublicKey,
+            ) -> Ciphered {
+                let msg = super::cbor_to_bytes(batch).unwrap();
+                let sig = sign_sk.sign(&msg);
+                let msg = Self {
+                    msg,
+                    sig,
+                    from,
+                    added_field: vec![127; 1024],
+                };
+                let msg = super::cbor_to_bytes(&msg).unwrap();
+                cipher_pk.encrypt(&msg, Self::ASSOCIATED_DATA).unwrap()
+            }
+
+            fn decrypt<T: DeserializeOwned>(
+                encrypted: &Ciphered,
+                cipher_sk: &hpke::SecretKey,
+            ) -> T {
+                let msg = cipher_sk.decrypt(encrypted, Self::ASSOCIATED_DATA).unwrap();
+                let Self { msg, .. } = super::cbor_from_bytes(&msg).unwrap();
+                super::cbor_from_bytes(&msg).unwrap()
+            }
+        }
+
+        #[derive(Debug, Serialize, Deserialize)]
+        enum NewMessage {
+            Triple(NewTripleMessage),
+            NewVariant(String),
+            #[serde(untagged)]
+            Unknown(ciborium::Value),
+        }
+
+        impl PartialEq<Message> for NewMessage {
+            fn eq(&self, other: &Message) -> bool {
+                match (self, other) {
+                    (NewMessage::Triple(a), Message::Triple(b)) => a == b,
+                    // ignore the unknowns for comparison since we don't care about them here.
+                    _ => true,
+                }
+            }
+        }
+
+        #[derive(Debug, Serialize, Deserialize)]
+        struct NewTripleMessage {
+            id: u64,
+            epoch: u64,
+            from: Participant,
+            #[serde(with = "serde_bytes")]
+            data: Vec<u8>,
+            timestamp: u64,
+            // added this new timestamp in the future:
+            #[serde(default)]
+            new_timestamp: Option<u64>,
+        }
+
+        impl PartialEq<TripleMessage> for NewTripleMessage {
+            fn eq(&self, other: &TripleMessage) -> bool {
+                self.id == other.id
+                    && self.epoch == other.epoch
+                    && self.from == other.from
+                    && self.data == other.data
+                    && self.timestamp == other.timestamp
+            }
+        }
+
+        let from = Participant::from(1337);
+        let (cipher_sk, cipher_pk) = mpc_keys::hpke::generate();
+        let sign_sk =
+            near_crypto::SecretKey::from_seed(near_crypto::KeyType::ED25519, "sign-encrypt1");
+        let mut participants = Participants::default();
+        participants.insert(
+            &from,
+            ParticipantInfo {
+                sign_pk: sign_sk.public_key(),
+                cipher_pk: cipher_pk.clone(),
+                id: from.into(),
+                url: "http://localhost:3030".to_string(),
+                account_id: "test.near".parse().unwrap(),
+            },
+        );
+        let participants = ParticipantMap::One(participants);
+
+        // Test forward compatibility
+        let old_batch = vec![
+            Message::Triple(TripleMessage {
+                id: 1234,
+                epoch: 0,
+                from,
+                data: vec![128; 1024],
+                timestamp: 1234567,
+            }),
+            Message::Generating(GeneratingMessage {
+                from,
+                data: vec![8; 512],
+            }),
+            Message::Signature(SignatureMessage {
+                id: SignId::new([7; 32]),
+                proposer: from,
+                presignature_id: 1234,
+                epoch: 0,
+                from,
+                data: vec![78; 1222],
+                timestamp: 1234567,
+            }),
+        ];
+        let encrypted = SignedMessage::encrypt(&old_batch, from, &sign_sk, &cipher_pk).unwrap();
+        let new_batch: Vec<NewMessage> = NewSignedMessage::decrypt(&encrypted, &cipher_sk);
+        assert_eq!(
+            new_batch, old_batch,
+            "encrypt/decrypt failed forward compatibility"
+        );
+
+        // Test backward compatibility
+        let new_batch = vec![
+            NewMessage::Triple(NewTripleMessage {
+                id: 1234,
+                epoch: 0,
+                from,
+                data: vec![128u8; 1024],
+                timestamp: 1234567,
+                new_timestamp: Some(777),
+            }),
+            NewMessage::NewVariant("hello".to_string()),
+        ];
+        let new_ciphered = NewSignedMessage::encrypt(&new_batch, from, &sign_sk, &cipher_pk);
+        let old_batch: Vec<Message> =
+            SignedMessage::decrypt(&new_ciphered, &cipher_sk, &participants).unwrap();
+        assert_eq!(
+            new_batch, old_batch,
+            "encrypt/decrypt failed backward compatibility"
+        );
+    }
+
+    #[test]
+    fn test_encrypt_size() {
+        let epoch = 1;
+        let from = Participant::from(0);
+        let batch = vec![
+            Message::Triple(TripleMessage {
+                id: 1,
+                epoch,
+                from,
+                data: vec![128u8; 1024],
+                timestamp: 1,
+            }),
+            Message::Triple(crate::protocol::message::TripleMessage {
+                id: 2,
+                epoch,
+                from,
+                data: vec![255u8; 2048],
+                timestamp: 2,
+            }),
+            Message::Triple(TripleMessage {
+                id: 3,
+                epoch,
+                from,
+                data: vec![101u8; 1337],
+                timestamp: 3,
+            }),
+        ];
+
+        let batch_bytesize = batch.iter().map(|msg| msg.size()).sum::<usize>();
+        dbg!(batch_bytesize);
+
+        let (_cipher_sk, cipher_pk) = hpke::generate();
+        let sign_sk =
+            near_crypto::SecretKey::from_seed(near_crypto::KeyType::ED25519, "sign-encrypt0");
+        let ciphered = SignedMessage::encrypt(&batch, from, &sign_sk, &cipher_pk).unwrap();
+        let ciphered_bytesize = ciphered.text.len();
+        dbg!(ciphered_bytesize);
+
+        let margin_percent = 0.05;
+        let margin_of_err = (batch_bytesize as f64 * margin_percent) as usize;
+        dbg!(margin_of_err);
+        assert!(
+            ((batch_bytesize - margin_of_err)..(batch_bytesize + margin_of_err))
+                .contains(&ciphered_bytesize),
+            "ciphered message size is not within 5% of the original message size"
+        );
+    }
+}
